--- a/src/polyphase/inducedcostheuristic.cpp
+++ b/src/polyphase/inducedcostheuristic.cpp
@@ -95,17 +95,14 @@ void InducedCostHeuristic::choosePermanentEdge(const StaticSparseGraph::Edge eIc
      * must be pairwise connected, too. For non-zero edges, we could let this be handled by the heuristic,
      * but for zero edges, we must do this.*/
     std::vector<Edge> implications;
-    std::vector<Edge> implicationsForbidden;
-    std::vector<NodeId> uClique(graph.getCliqueOf(eIcf.u));
-    std::vector<NodeId> vClique(graph.getCliqueOf(eIcf.v));
     
     /* All pairs from uClique and vClique are found. We must be careful not to add eIcf to our list (we already have it) and
      * to not add edges, which are already permanent. We delay the actual modification of the edge until we know, which edges
      * must become permanent, because the weight for zero-edges implicitly changes, if other edges change. This would disturb
      * the search process.*/
     //TODO: Find reason, why some edges are already permanent here.
-    for (NodeId x : uClique) {
-        for (NodeId y : vClique) {
+    for (NodeId x : graph.getCliqueOf(eIcf.u)) {
+        for (NodeId y : graph.getCliqueOf(eIcf.v)) {
             Edge e = Edge(x,y);
             if (x == y || graph.findIndex(e) == 0 || (x == eIcf.u && y == eIcf.v)) {
                 continue;
@@ -116,8 +113,11 @@ void InducedCostHeuristic::choosePermanentEdge(const StaticSparseGraph::Edge eIc
     
     /* The cliques we are connecting here might already be forbidden to other nodes/cliques. So, we need a second list of
      * implications to collect edges, which must be set to forbidden afterwards.*/
+    std::vector<Edge> implicationsForbidden;
     for (NodeId f : graph.getForbiddenNeighbors(eIcf.u)) {
-        for (NodeId x : vClique) {
+        if (graph.isForbidden(Edge(eIcf.v, f)))
+            continue;
+        for (NodeId x : graph.getCliqueOf(eIcf.v)) {
             Edge e = Edge(f,x);
             if (graph.findIndex(e) != 0 && !graph.isForbidden(e)) {
                 implicationsForbidden.push_back(e);
@@ -125,7 +125,9 @@ void InducedCostHeuristic::choosePermanentEdge(const StaticSparseGraph::Edge eIc
         }
     }
     for (NodeId f : graph.getForbiddenNeighbors(eIcf.v)) {
-        for (NodeId x : uClique) {
+        if (graph.isForbidden(Edge(eIcf.u, f)))
+            continue;
+        for (NodeId x : graph.getCliqueOf(eIcf.u)) {
             Edge e = Edge(f,x);
             if (graph.findIndex(e) != 0 && !graph.isForbidden(e)) {
                 implicationsForbidden.push_back(e);
@@ -158,6 +160,9 @@ void InducedCostHeuristic::choosePermanentEdge(const StaticSparseGraph::Edge eIc
         /* Idea: We iterate over all outgoing edges from the combined clique of u and v. If we reach another cluster, we have not seen
          * before, we memorize the outgoing edge as representative edge for this cluster. When we find another edge leading to a cluster,
          * we have already seen, we merge this edge with the representative of this cluster.*/
+        
+        std::vector<NodeId> uClique(graph.getCliqueOf(eIcf.u));
+        std::vector<NodeId> vClique(graph.getCliqueOf(eIcf.v));
         uClique.insert(uClique.end(), vClique.begin(), vClique.end());
         for (NodeId x : uClique) {
             for (NodeId xn : graph.getUnprunedNeighbours(x)) {
@@ -188,16 +193,14 @@ void InducedCostHeuristic::chooseForbiddenEdge(const StaticSparseGraph::Edge eIc
      * all other pair of nodes in the same clique as u and v must be a forbidden pair. For non-zero edges, we could let 
      * this be handled by the heuristic, but for zero edges, we must do this.*/
     std::vector<Edge> implications;
-    std::vector<NodeId> uClique(graph.getCliqueOf(eIcp.u));
-    std::vector<NodeId> vClique(graph.getCliqueOf(eIcp.v));
     
     /* All pairs from uClique and vClique are found. We must be careful not to add eIcp to our list (we already have it) and
      * to not add edges, which are already forbidden. We delay the actual modification of the edge until we know, which edges
      * must become forbidden, because the weight for zero-edges implicitly changes, if other edges change. This would disturb
      * the search process.*/
     //TODO: Find reason, why some edges are already forbidden here.
-    for (NodeId x : uClique) {
-        for (NodeId y : vClique) {
+    for (NodeId x : graph.getCliqueOf(eIcp.u)) {
+        for (NodeId y : graph.getCliqueOf(eIcp.v)) {
             Edge e = Edge(x,y);
             if (x == y || graph.findIndex(e) == 0 || (x == eIcp.u && y == eIcp.v)) {
                 continue;
@@ -366,23 +369,19 @@ void InducedCostHeuristic::setPermanent(const Edge e) {
 }
 
 void InducedCostHeuristic::updateTripleForbiddenUW(const EdgeWeight uv, const Edge uw, const EdgeWeight vw) {
-    EdgeWeight icf_old = edgeHeap.getIcf(uv, vw);
-    EdgeWeight icf_new = 0.0;
-    EdgeWeight icp_old = edgeHeap.getIcp(uv, vw);
-    EdgeWeight icp_new = std::max(0.0f, vw);
-    if (icf_new != icf_old)
-        edgeHeap.increaseIcf(uw, icf_new - icf_old);
-    if (icp_new != icp_old)
-        edgeHeap.increaseIcp(uw, icp_new - icp_old);
+    EdgeWeight icfChange = 0.0 - edgeHeap.getIcf(uv, vw);
+    EdgeWeight icpChange = std::max(0.0f, vw) - edgeHeap.getIcp(uv, vw);
+    if (icfChange != 0)
+        edgeHeap.increaseIcf(uw, icfChange);
+    if (icpChange != 0)
+        edgeHeap.increaseIcp(uw, icpChange);
 }
 
 void InducedCostHeuristic::updateTriplePermanentUW(const EdgeWeight uv, const Edge uw, const EdgeWeight vw) {
-    EdgeWeight icf_old = edgeHeap.getIcf(uv, vw);
-    EdgeWeight icf_new = std::max(0.0f, vw);
-    EdgeWeight icp_old = edgeHeap.getIcp(uv, vw);
-    EdgeWeight icp_new = std::max(0.0f, -vw);
-    if (icf_new != icf_old)
-        edgeHeap.increaseIcf(uw, icf_new - icf_old);
-    if (icp_new != icp_old)
-        edgeHeap.increaseIcp(uw, icp_new - icp_old);
+    EdgeWeight icfChange = std::max(0.0f, vw) - edgeHeap.getIcf(uv, vw);
+    EdgeWeight icpChange = std::max(0.0f, -vw) - edgeHeap.getIcp(uv, vw);
+    if (icfChange != 0)
+        edgeHeap.increaseIcf(uw, icfChange);
+    if (icpChange != 0)
+        edgeHeap.increaseIcp(uw, icpChange);
 }