… tests/test_clusterediting.py