I found a nice shortcut in the cluster editing heuristic, which does not change the results (or at least it should not and I did not find any differences so far), but makes the heuristic run in less than a third of the old time. This makes the regular polyphase algorithm 20-30% faster. The upcoming genetic polyphase should profit much more, because the cluster editing is much more of a bottleneck there, at least on large instances. I also added/updates the testcases for cluster editing.

I did not edit the `CHANGES.rst` yet, because I first wanted to wait for approval.