@marcelm 2022-06-01
Awesome! Do you have anyone in particular in mind who can approve this? I’m not familiar with the code, so I cannot review it. I think if the tests pass, it should go in after adding a changelog entry as you said.