@schrins 2022-06-05
There is probably no one besides me who knows the cluster editing C++ code. I added a few more changes, most of them actually were in the main branch for some reason (or I should rather say: they were not in the feature branch for some reason) and tested some instances again to make sure the output phasings are identical. If there are no objections, I think we should merge this branch.

Speed-ups are hard to measure (see below), so I stayed with a conservative 20-30% speed-up in the changelog.
For a sample of the tetraploid human data from the polyphase paper, the cluster editing time went down from 72 to 29 seconds and the total time went down from 198 to 155 seconds. The extrem case is a sample of the test data provided by another polyphase algorithm (Flopp by Shaw et al. 2022), because the reads form extremely big clusters: The clustering time went down from 63 to 2 (!) seconds with overall time going down from 114 to 52 seconds.